cleaned up prints
