+++ processor/__init__.py
@@ -1,2 +1,3 @@
 from .lambda_function import handler
 from .lambda_function import get_json
+from .lambda_function import process_qandl_data
